requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.

#-r prod.txt

# Documentation
#Sphinx==1.0.6

# Testing
#nose==1.0.0
#-e git://github.com/jbalogh/django-nose.git#egg=django_nose
#-e git://github.com/jbalogh/test-utils.git#egg=test-utils

# L10n
#translate-toolkit==1.8.0

# Mocking
mock

# for making migrations
django-extensions

-e git+https://github.com/jbalogh/check.git#egg=check
